+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.2.0]
 
+### Breaking Changes
+
+- Renamed all methods from `delete` to `remove` and all of them returning `AffectedRecords`
+- Renamed all methods from `deleteFields` to `removeFields`
+
 ### Added
 
 - Localizations Service
@@ -21,8 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed return types of `sdk.users.getStaff` and `sdk.users.getPatients`
 - Updated docs links
-- Gzip option `sdk.data.documents.create`
 - `updateProfileImage` method is deprecated
+- Gzip option `sdk.data.documents.create`
 
 ```ts
 await sdk.data.documents.create('schemaId', document, {